refactor map buttons
