+++ lib/widgets/map/buttons/crop_button.dart
